+++ app/src/main/java/com/example/hsinhwang/shrimpshell/CustomerPanel/ProfileInformationFragment.java
@@ -1,4 +1,4 @@
-package com.example.hsinhwang.shrimpshell;
+package com.example.hsinhwang.shrimpshell.CustomerPanel;
 
 import android.Manifest;
 import android.content.Intent;
@@ -16,6 +16,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+
+import com.example.hsinhwang.shrimpshell.Classes.Common;
+import com.example.hsinhwang.shrimpshell.R;
 
 public class ProfileInformationFragment extends Fragment{
     private static final int RESULT_OK = -1;
