+++ app/src/main/java/com/example/hsinhwang/shrimpshell/CustomerPanel/ProfileReceiptFragment.java
@@ -1,4 +1,4 @@
-package com.example.hsinhwang.shrimpshell;
+package com.example.hsinhwang.shrimpshell.CustomerPanel;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -9,6 +9,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.example.hsinhwang.shrimpshell.R;
 
 import java.util.ArrayList;
 import java.util.List;
