+++ app/src/main/res/layout/activity_event.xml
@@ -7,8 +7,8 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".EventActivity"
-    android:background="@drawable/employee_home">
+    tools:context=".GeneralPages.EventActivity"
+    android:background="@drawable/customer_global_background">
 
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
