+++ app/src/main/res/layout/activity_introduction.xml
@@ -11,13 +11,14 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="飯店介紹"
-        android:textSize="24dp"
+        android:textSize="24sp"
+        android:textColor="#FFFFFF"
         android:layout_marginBottom="10dp"/>
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/introduction1"
-        android:textSize="22dp"
+        android:textSize="22sp"
         android:lineSpacingExtra="15dp"
         android:textColor="#FFFFFF"/>
 
@@ -25,7 +26,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/introduction2"
-        android:textSize="22dp"
+        android:textSize="22sp"
         android:lineSpacingExtra="15dp"
         android:layout_marginTop="16dp"
         android:textColor="#FFFFFF"
