+++ app/src/main/res/layout/activity_room_detail.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:paddingHorizontal="16dp"
-    tools:context=".EventActivity">
+    tools:context=".GeneralPages.EventActivity">
 
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
