+++ app/src/main/res/values/strings.xml
@@ -44,6 +44,11 @@
     <string name="add">Add</string>
     <string name="male">Male</string>
     <string name="female">Female</string>
+    <string name="lastest_events">Latest Events</string>
+    <string name="rooms">Rooms</string>
+    <string name="location">Location</string>
+    <string name="login">Account Login</string>
+    <string name="about">About Us</string>
 
 </resources>
